Lights still busted
Fixed a method name, though lights are still broken. Output of the
brightness values shows everything is negative. However, it's going to
be because we're calculating the FFT of the raw data and not the actual
audio frames from the song. Need to be able to read frames AND chunk
data. Easiest answer is probably pre-cache the file.