Separates the audio input and output
Moves audio input and output into their own modules and provides
adapters so the audio-in path is equivalent to a file input path, at
least in theory. I haven't yet tested the microphone/line-in portion,
but it *should* work the same. More work needs to be done as I want to
have a conditional cache middleware in the pipeline that either fetches
the matrix from cache or generates it live if we couldn't pre-cache the
file.