New playlist class and playlist playthrough
Adds a new Playlist class to represent the list of songs to be played,
and implements real shuffling of the list using random.shuffle so the
same song is never played twice in a row. Additionally, modifies
synchronized_lights.py to be the playlist player, rather than simply
playing one song and exiting.

Does remove the voting functionality and many of the related bits for
the time being, but I didn't intend to use that any time soon anyway.